+++ examples/android-instrumentation-sample/src/main/java/io/sentry/samples/instrumentation/ui/MainActivity.kt
@@ -20,45 +20,48 @@ import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.withContext
 
 class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+  override fun onCreate(savedInstanceState: Bundle?) {
+    super.onCreate(savedInstanceState)
+    setContentView(R.layout.activity_main)
 
-        val list = findViewById<RecyclerView>(R.id.list)
-        list.layoutManager = LinearLayoutManager(this)
-        list.adapter = TrackAdapter()
+    val list = findViewById<RecyclerView>(R.id.list)
+    list.layoutManager = LinearLayoutManager(this)
+    list.adapter = TrackAdapter()
 
-        lifecycleScope.launchWhenStarted {
-            Sentry.getSpan()?.finish()
-            val transaction = Sentry.startTransaction(
-                "Track Interaction",
-                "ui.action.load",
-                TransactionOptions().apply { isBindToScope = true }
-            )
-            SampleApp.database.tracksDao()
-                .all()
-                .map {
-                    val remote = withContext(Dispatchers.IO) {
-                        TrackService.instance.tracks("6188aa82-3102-436a-9a68-513e6ad9efcb")
-                    }
-                    remote + it
-                }
-                .collect {
-                    (list.adapter as TrackAdapter).populate(it)
-                    transaction.finish(SpanStatus.OK)
-                }
+    lifecycleScope.launchWhenStarted {
+      Sentry.getSpan()?.finish()
+      val transaction =
+        Sentry.startTransaction(
+          "Track Interaction",
+          "ui.action.load",
+          TransactionOptions().apply { isBindToScope = true },
+        )
+      SampleApp.database
+        .tracksDao()
+        .all()
+        .map {
+          val remote =
+            withContext(Dispatchers.IO) {
+              TrackService.instance.tracks("6188aa82-3102-436a-9a68-513e6ad9efcb")
+            }
+          remote + it
         }
-
-        findViewById<Toolbar>(R.id.toolbar).setOnMenuItemClickListener {
-            if (it.itemId == R.id.action_add) {
-                startActivity(Intent(this, EditActivity::class.java))
-                return@setOnMenuItemClickListener true
-            }
-            if (it.itemId == R.id.action_compose) {
-                startActivity(Intent(this, ComposeActivity::class.java))
-                return@setOnMenuItemClickListener true
-            }
-            return@setOnMenuItemClickListener false
+        .collect {
+          (list.adapter as TrackAdapter).populate(it)
+          transaction.finish(SpanStatus.OK)
         }
     }
+
+    findViewById<Toolbar>(R.id.toolbar).setOnMenuItemClickListener {
+      if (it.itemId == R.id.action_add) {
+        startActivity(Intent(this, EditActivity::class.java))
+        return@setOnMenuItemClickListener true
+      }
+      if (it.itemId == R.id.action_compose) {
+        startActivity(Intent(this, ComposeActivity::class.java))
+        return@setOnMenuItemClickListener true
+      }
+      return@setOnMenuItemClickListener false
+    }
+  }
 }